# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delicious Dinner API",
        "description": "API realizada en Laravel",
        "contact": {
            "email": "ingisraeltrujillo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/categoria_producto": {
            "get": {
                "tags": [
                    "categoria_producto"
                ],
                "summary": "Muestra todas las categorias de productos",
                "description": "Muestra todas las categorias de productos.",
                "operationId": "298d86f2eb9eed4a60558a948061b7c7",
                "responses": {
                    "200": {
                        "description": "Muestra todas las categorias de productos"
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "categoria_producto"
                ],
                "summary": "Crear una nueva categoria de producto",
                "operationId": "8db91566178205ae0db6dbef4172de7e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "codename": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "HAMBURGUESA",
                                    "codename": "burger"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED"
                    }
                }
            }
        },
        "/api/categoria_producto/{id}": {
            "get": {
                "tags": [
                    "categoria_producto"
                ],
                "summary": "Mostrar la información de una categoria de producto",
                "operationId": "1359e55bba9e81066e9277a00c16a596",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "categoria_producto"
                ],
                "summary": "Actualizar una categoria de producto",
                "operationId": "9ccecf5132ea3d184ade0840270a3121",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "codename": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "HAMBURGUESA",
                                    "codename": "burger"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "categoria_producto"
                ],
                "summary": "Eliminar una categoria de producto",
                "operationId": "e63fabf1fd95ae992710f10876b23ee0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
...
    }
}